backend/requirements.txt
fastapi
uvicorn
cryptography
python-jose[cryptography]
passlib[bcrypt]
ccxt
protobuf>=5.29.3
